fix(Table): Remove focus from `Table` with interactive elements
