+++ .changeset/fix-Table-remove-focus.md
@@ -1,0 +1,6 @@
+---
+
+'react-magma-dom': patch
+---
+
+fix(Table): Remove focus from `Table` with interactive elements
